refactor(auth-service): split main.ts into two
keep the routing stuff inside app.ts
Setup stuff inside main.ts
